// src/api/__generated__/mining/model/weeklyRewardsDTO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */

/**
 * 每周奖励响应
 */
export interface WeeklyRewardsDTO {
  /** 奖励周期开始时间 */
  rewardPeriodStart?: string;
  /** 奖励周期结束时间 */
  rewardPeriodEnd?: string;
  /**
   * 周硬顶
   * @minimum 0
   */
  hardLimit?: number;
  /**
   * 质押挖矿STORY数量
   * @minimum 0
   */
  miningRewards?: number;
  /**
   * 邀请奖励STORY数量
   * @minimum 0
   */
  inviteRewards?: number;
}